Had to reconfigure the switches
We cannot use the switch configured as a tamper pin. According to my additional reading, that is to indicate if someone has attempted to tamper with the device and the MCU will detect this and delete data as a result. We will use SW1 as an external interrupt. This configures SW1 as EXTI12 and also configures the Nucleo's SW2 as EXTI0 for testing purposes.