PVD is working, removed extra app_ble.c and p2p_server_app.c files, changed advertising to low power mode
The PVD is tripped at 2.8 V and the interrupt handler then resets the level to 2.6 V. Right now the LEDs are just toggling so that will need to be updated.

Found the extra files, not sure how they appeared but they are gone.

Changed advertising mode to LP (low power) - now that the PVD is mostly implemented I am going to focus on LP until FRAM implementation